requirements.txt
Django~=4.0.3
djangorestframework~=3.13.1
selenium~=4.1.5